test: payment challenge flow
