+++ graphql/schemas/Souk/Payments/payments.graphql
@@ -126,7 +126,7 @@ extend type Mutation @guard {
         @field(
             resolver: "App\\GraphQL\\Souk\\Mutations\\Payments\\PaymentMutation@completeDeviceData"
         )
-    validatePayerAuthResult(orderId: ID!, transactionId: String!): PayerAuthStatus!
+    validatePayerAuthResult(orderId: ID!): PayerAuthStatus!
         @field(
             resolver: "App\\GraphQL\\Souk\\Mutations\\Payments\\PaymentMutation@validatePayerAuthResult"
         )
