+++ src/Domains/Connectors/EchoPay/Enums/CustomFieldEnum.php
@@ -11,7 +11,7 @@ enum CustomFieldEnum: string
     case ECHO_PAY_PAYMENT_STATUS = 'payment_status';
     case ECHO_PAY_PAYMENT_RESPONSE = 'payment_response';
     case ECHO_PAY_SHOULD_CAPTURE = 'should_capture';
-
+    case ECHO_PAY_AUTH_TRANSACTION_ID = 'auth_transaction_id';
 
     case ECHO_PAY_MERCHANT_KEY = 'merchant_key';
     case ECHO_PAY_CHANNEL_CODE = 'channel_code';
